requirements.txt
asgiref==3.5.2
certifi==2021.10.8
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
itypes==1.2.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.3
webencodings==0.5.1
xlwt==1.3.0
